igzip: prevent internal buffer overrun
In fixed_size_read(), the condition that decides whether to buffer
incoming data into tmp_in_buffer uses a 32-bit addition:

  state->avail_in + tmp_in_size < (uint32_t) read_size

If tmp_in_size is nonzero and state->avail_in is close to UINT32_MAX,
the sum wraps around, producing a value smaller than read_size. This
incorrectly takes the buffering branch and copies state->avail_in bytes
(~4 GB) into the fixed-size tmp_in_buffer, causing a heap buffer
overflow write.

+++ igzip/igzip_inflate.c
@@ -1766,11 +1766,11 @@ isal_inflate_reset(struct inflate_state *state)
 }
 
 static inline uint32_t
-fixed_size_read(struct inflate_state *state, uint8_t **read_buf, int read_size)
+fixed_size_read(struct inflate_state *state, uint8_t **read_buf, uint32_t read_size)
 {
         uint32_t tmp_in_size = state->tmp_in_size;
 
-        if (state->avail_in + tmp_in_size < (uint32_t) read_size) {
+        if ((uint64_t) state->avail_in + tmp_in_size < read_size) {
                 memcpy(state->tmp_in_buffer + tmp_in_size, state->next_in, state->avail_in);
                 tmp_in_size += state->avail_in;
                 state->tmp_in_size = tmp_in_size;
